remove dependency on otp and adapt to use r5 directly
previously, analyst-server depended on otp, which in turn depended on r5.
we are switching over to do all analysis using r5.
this change was made now because new and different scenario modification classes were being added to r5, so the external interface of the two systems was beginning to drift apart.